# [dtrace-perf-map] Use binary rather than linear search